Kernel: Fix build warnings in Alpine Linux
Alpine uses musl instead of glibc, which has incorrectly
defined uint64_t as long, which on -m32 is 32-bit

Since I only use uint64_t in one place, for now this workaround
is good enough

+++ include/kernel.h
@@ -28,8 +28,8 @@ typedef void (*isr_handler_fn)(isr_stack_st *isr_stack);
 
 typedef struct {
     uint32_t size;
-    uint64_t addr;
-    uint64_t len;
+    unsigned long long addr;
+    unsigned long long len;
     uint32_t type;
 } __attribute__((packed)) mboot_mmap_entry_st;
